fix InitGlobalEnvironment
