+++ src/Environment/Global/InitGnssSatellites.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <Environment/Global/GnssSatellites.h>
+
+GnssSatellites* InitGnssSatellites(std::string file_name);
